Add pytiger.utils.plugins
This is a simple plugin loading mechanism: it provides a function that
can load Python modules and packages from a given directory, returning
a list of module objects corresponding to the loaded modules. The
loading code may either inspect the modules through the objects in this
list, or the plugins themselves may register themselves with the
loading application by unspecified means.