Add GitLab CI configuration
We run an internal GitLab instance which we can use to run CI jobs as
well as (or instead of) Travis CI if needs be, so configure some jobs.
This also allows us to test on Python 3.7 which Travis CI currently
doesn't support[1], and we can also build the documentation and host it
in GitLab Pages.

1. https://github.com/travis-ci/travis-ci/issues/9815